Move Random to new package of same name
