+++ app/src/main/java/Bamboo/controller/random/Random.java
@@ -1,5 +1,7 @@
-package Bamboo.controller;
+package Bamboo.controller.random;
 
+import Bamboo.controller.Agent;
+import Bamboo.controller.Vector;
 import Bamboo.model.Game;
 import Bamboo.model.Tile;
